Move plotly functions to scripts folder.
